# Conflicts:
#	app/src/main/java/org/dhis2/data/metadata/MetadataRepositoryImpl.java
#	app/src/main/java/org/dhis2/usescases/datasets/datasetDetail/DataSetDetailRepositoryImpl.java
#	app/src/main/java/org/dhis2/usescases/datasets/datasetInitial/DataSetInitialActivity.java
#	app/src/main/java/org/dhis2/usescases/main/MainPresenter.java
#	app/src/main/java/org/dhis2/usescases/main/program/HomeRepositoryImpl.java
#	app/src/main/java/org/dhis2/usescases/searchTrackEntity/adapters/TabletSearchAdapter.java
#	app/src/main/java/org/dhis2/utils/Constants.java
#	app/src/main/java/org/dhis2/utils/DateUtils.java
#	app/src/main/java/org/dhis2/utils/custom_views/DateTimeView.java
#	app/src/main/java/org/dhis2/utils/custom_views/DateView.java
#	app/src/main/java/org/dhis2/utils/custom_views/TimeView.java
#	app/src/main/res/values/strings.xml
#	buildsystem/dependencies.gradle
#	settings.gradle

+++ app/src/main/java/org/dhis2/data/forms/FormFragment.java
@@ -43,6 +43,7 @@ import org.hisp.dhis.rules.models.RuleActionErrorOnCompletion;
 import org.hisp.dhis.rules.models.RuleActionShowError;
 import org.hisp.dhis.rules.models.RuleActionWarningOnCompletion;
 
+import java.util.Date;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
@@ -103,6 +104,10 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
     private RuleActionShowError showError;
     private String programUid;
     private String teiUid;
+    private Date openingDate;
+    private Date closingDate;
+    private boolean mandatoryDelete = true;
+    private Context context;
 
 
     public FormFragment() {
@@ -240,6 +245,7 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
+        this.context = context;
         FormViewArguments arguments = Preconditions.isNull(getArguments()
                 .getParcelable(FORM_VIEW_ARGUMENTS), "formViewArguments == null");
 
@@ -252,6 +258,7 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
 
     @Override
     public void onDetach() {
+        context = null;
         super.onDetach();
     }
 
@@ -323,6 +330,12 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
         return captureCoordinates -> coordinatesView.setVisibility(captureCoordinates ? View.VISIBLE : View.GONE);
     }
 
+    @Override
+    public void setMinMaxDates(Date openingDate, Date closingDate) {
+        this.openingDate = openingDate;
+        this.closingDate = closingDate;
+    }
+
     @NonNull
     @Override
     public Consumer<String> renderTitle() {
@@ -377,12 +390,14 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
     public void initReportDatePicker(boolean reportAllowFutureDates, boolean incidentAllowFutureDates) {
         reportDate.setOnClickListener(v -> {
             DatePickerDialogFragment dialog = DatePickerDialogFragment.create(reportAllowFutureDates);
+            dialog.setOpeningClosingDates(openingDate, closingDate);
             dialog.show(getFragmentManager());
             dialog.setFormattedOnDateSetListener(publishReportDateChange());
         });
 
         incidentDate.setOnClickListener(v -> {
             DatePickerDialogFragment dialog = DatePickerDialogFragment.create(incidentAllowFutureDates);
+            dialog.setOpeningClosingDates(openingDate, closingDate);
             dialog.show(getFragmentManager());
             dialog.setFormattedOnDateSetListener(publishIncidentDateChange());
         });
@@ -390,6 +405,7 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
         reportDate.setOnFocusChangeListener((v, hasFocus) -> {
             if (hasFocus) {
                 DatePickerDialogFragment dialog = DatePickerDialogFragment.create(reportAllowFutureDates);
+                dialog.setOpeningClosingDates(openingDate, closingDate);
                 dialog.show(getFragmentManager());
                 dialog.setFormattedOnDateSetListener(publishReportDateChange());
             }
@@ -398,6 +414,7 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
         incidentDate.setOnFocusChangeListener((v, hasFocus) -> {
             if (hasFocus) {
                 DatePickerDialogFragment dialog = DatePickerDialogFragment.create(incidentAllowFutureDates);
+                dialog.setOpeningClosingDates(openingDate, closingDate);
                 dialog.show(getFragmentManager());
                 dialog.setFormattedOnDateSetListener(publishIncidentDateChange());
             }
@@ -437,21 +454,28 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         switch (requestCode) {
             case Constants.RQ_MAP_LOCATION_VIEW:
-                coordinatesView.updateLocation(Double.valueOf(data.getStringExtra(MapSelectorActivity.LATITUDE)), Double.valueOf(data.getStringExtra(MapSelectorActivity.LONGITUDE)));
-                publishCoordinatesChanged(Double.valueOf(data.getStringExtra(MapSelectorActivity.LATITUDE)), Double.valueOf(data.getStringExtra(MapSelectorActivity.LONGITUDE)));
-                this.coordinatesView = null;
+                if (data != null && data.getStringExtra(MapSelectorActivity.LATITUDE) != null && data.getStringExtra(MapSelectorActivity.LONGITUDE) != null) {
+                    coordinatesView.updateLocation(Double.valueOf(data.getStringExtra(MapSelectorActivity.LATITUDE)), Double.valueOf(data.getStringExtra(MapSelectorActivity.LONGITUDE)));
+                    publishCoordinatesChanged(Double.valueOf(data.getStringExtra(MapSelectorActivity.LATITUDE)), Double.valueOf(data.getStringExtra(MapSelectorActivity.LONGITUDE)));
+                    this.coordinatesView = null;
+                }
                 break;
             case RQ_EVENT:
-                openDashboard();
+                if (data != null)
+                    openDashboard(data.getStringExtra(Constants.EVENT_UID));
+                break;
+            default:
                 break;
         }
         super.onActivityResult(requestCode, resultCode, data);
     }
 
-    void openDashboard(){
+    private void openDashboard(@Nullable String eventUid) {
         Bundle bundle = new Bundle();
         bundle.putString("PROGRAM_UID", programUid);
         bundle.putString("TEI_UID", teiUid);
+        if (eventUid != null)
+            bundle.putString(Constants.EVENT_UID, eventUid);
         startActivity(TeiDashboardMobileActivity.class, bundle, false, false, null);
         getActivity().finish();
     }
@@ -493,12 +517,10 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
                     eventCreationIntent.putExtras(EventCaptureActivity.getActivityBundle(enrollmentTrio.val2(), enrollmentTrio.val1()));
                     eventCreationIntent.putExtra(Constants.TRACKED_ENTITY_INSTANCE, enrollmentTrio.val0());
                     startActivityForResult(eventCreationIntent, RQ_EVENT);
-                    /*FormViewArguments formViewArguments = FormViewArguments.createForEvent(enrollmentTrio.val2());
-                    startActivityForResult(FormActivity.create(getContext(), formViewArguments, isEnrollment), RQ_EVENT);*/
                 } else { //val0 is program uid, val1 is trackedEntityInstance, val2 is empty
                     this.programUid = enrollmentTrio.val1();
                     this.teiUid = enrollmentTrio.val0();
-                    openDashboard();
+                    openDashboard(null);
                 }
             } else {
                 checkAction();
@@ -509,48 +531,61 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
     }
 
     private void checkAction() {
-        CustomDialog dialog = new CustomDialog(
-                getContext(),
-                getString(R.string.warning_error_on_complete_title),
-                messageOnComplete,
-                getString(R.string.button_ok),
-                getString(R.string.cancel),
-                1001,
-                new DialogClickListener() {
-                    @Override
-                    public void onPositive() {
-                        if (canComplete)
-                            getActivity().finish();
-                    }
+        if (isAdded() && getContext() != null) {
+            CustomDialog dialog = new CustomDialog(
+                    getContext(),
+                    getString(R.string.warning_error_on_complete_title),
+                    messageOnComplete,
+                    getString(R.string.button_ok),
+                    getString(R.string.cancel),
+                    1001,
+                    new DialogClickListener() {
+                        @Override
+                        public void onPositive() {
+                            if (canComplete && getActivity() != null && isAdded())
+                                getActivity().finish();
+                        }
 
-                    @Override
-                    public void onNegative() {
-                    }
-                });
-        if (!isEmpty(messageOnComplete))
-            dialog.show();
-        else
-            getActivity().finish();
+                        @Override
+                        public void onNegative() {
+                        }
+                    });
+            if (isAdded() && !isEmpty(messageOnComplete) && !dialog.isShowing())
+                dialog.show();
+            else if (isAdded() && getActivity() != null) {
+                getActivity().finish();
+            }
+        }
     }
 
     @Override
     public void showMandatoryFieldsDialog() {
+        String description;
+        String buttonAccept;
+        if (mandatoryDelete) {
+            description = isEnrollment ? getAbstracContext().getString(R.string.missing_mandatory_fields_text) :
+                    getAbstracContext().getString(R.string.missing_mandatory_fields_events);
+            buttonAccept = isEnrollment ? getAbstracContext().getString(R.string.missing_mandatory_fields_go_back) :
+                    getAbstracContext().getString(R.string.button_ok);
+        } else {
+            description = getAbstracContext().getString(R.string.missing_mandatory_fields_text_not_delete);
+            buttonAccept = getAbstracContext().getString(R.string.action_accept);
+        }
         new CustomDialog(
                 getAbstracContext(),
                 getAbstracContext().getString(R.string.missing_mandatory_fields_title),
-                isEnrollment ? getAbstracContext().getString(R.string.missing_mandatory_fields_text) :
-                        getAbstracContext().getString(R.string.missing_mandatory_fields_events),
-                isEnrollment ? getAbstracContext().getString(R.string.missing_mandatory_fields_go_back) :
-                        getAbstracContext().getString(R.string.button_ok),
+                description,
+                buttonAccept,
                 getAbstracContext().getString(R.string.cancel),
                 RC_GO_BACK,
                 new DialogClickListener() {
                     @Override
                     public void onPositive() {
-                        if (isEnrollment)
-                            deleteAllSavedDataAndGoBack();
-                        else
-                            getActivity().finish();
+                        if (mandatoryDelete)
+                            if (isEnrollment)
+                                deleteAllSavedDataAndGoBack();
+                            else
+                                getActivity().finish();
                     }
 
                     @Override
@@ -573,6 +608,11 @@ public class FormFragment extends FragmentGlobalAbstract implements FormView, Co
 
     @Override
     public void onBackPressed() {
+        formPresenter.checkMandatoryFields();
+    }
+
+    public void onBackPressed(boolean delete) {
+        this.mandatoryDelete = delete;
         formPresenter.checkMandatoryFields();
     }
